pmg/rpc: make freespace and height u64
when using a restricted port, freespace gives back u64's max value[1]:

	std::numeric_limits<uint64_t>::max()

[1]: https://github.com/monero-project/monero/blob/ed506006d2563c474e52db5a91ba2e20af86f30f/src/rpc/core_rpc_server.cpp#L511

+++ pkg/rpc/daemon/types.go
@@ -210,9 +210,9 @@ type GetInfoResult struct {
 	BusySyncing              bool   `json:"busy_syncing"`
 	CumulativeDifficulty     int64  `json:"cumulative_difficulty"`
 	Difficulty               int64  `json:"difficulty"`
-	FreeSpace                int64  `json:"free_space"`
+	FreeSpace                uint64 `json:"free_space"`
 	GreyPeerlistSize         int    `json:"grey_peerlist_size"`
-	Height                   int    `json:"height"`
+	Height                   uint64 `json:"height"`
 	HeightWithoutBootstrap   int    `json:"height_without_bootstrap"`
 	IncomingConnectionsCount int    `json:"incoming_connections_count"`
 	Mainnet                  bool   `json:"mainnet"`
